chore: adjust apply test file names and imports
