+++ test/unit/engine/apply-changeset/tasks/create-change-entities-task.test.ts
@@ -1,25 +1,29 @@
 import { initializeTask } from '../../../test-utils'
 import { expect } from 'chai'
-import { applyAddEntitiesTask } from '../../../../../src/engine/apply-changeset/tasks/create-add-entities-taskk'
 import { beforeEach } from 'mocha'
-import { AddedChangesetItem } from '../../../../../src/engine/types'
+import { UpdatedChangesetItem } from '../../../../../src/engine/types'
 import { createLinkObject } from '../../../../../src/engine/utils/create-link-object'
 import { createApplyChangesetContext } from '../../../fixtures/apply-changeset-context-fixture'
 import { ApplyChangesetContext } from '../../../../../src/engine/apply-changeset/types'
+import { ApplyChangesetTasks } from '../../../../../src/engine/apply-changeset/tasks'
 
-describe('applyAddEntitiesTask', () => {
+describe('applyChangeEntitiesTask', () => {
   let context: ApplyChangesetContext
   beforeEach(() => {
     context = createApplyChangesetContext()
   })
 
-  it('creates an entry from changeset', async () => {
-    const addedChangesetItem: AddedChangesetItem = createLinkObject('added-entry', 'add', 'Entry')
+  it('updates an entry from changeset', async () => {
+    const updatedChangesetItem: UpdatedChangesetItem = {
+      ...createLinkObject('update-entry', 'update', 'Entry'),
+      patch: [],
+    }
 
-    addedChangesetItem.data = {
+    const updatedEntry = {
+      metadata: { tags: [] },
       sys: {
         space: { sys: { type: 'Link', linkType: 'Space', id: '529ziq3ce86u' } },
-        id: 'added-entry',
+        id: 'update-entry',
         type: 'Entry',
         createdAt: '2023-05-17T10:36:22.538Z',
         updatedAt: '2023-05-17T10:36:40.280Z',
@@ -30,32 +34,27 @@ describe('applyAddEntitiesTask', () => {
       fields: {},
     }
 
-    const createdEntry = {
-      sys: {
-        id: 'added-entry',
-      },
-    }
-
     class Spy {
       called = 0
 
       call = (args: any): any => {
         expect(args.environment).to.be.equal('qa')
-        expect(args.entryId).to.be.equal('added-entry')
-        expect(args.contentType).to.be.equal('lesson')
+        expect(args.entryId).to.be.equal('update-entry')
 
         this.called++
 
-        return createdEntry
+        return updatedEntry
       }
     }
 
     const spy = new Spy()
 
-    context.client.cma.entries.create = spy.call
-    context.changeset.items.push(addedChangesetItem)
+    context.client.cma.entries.update = spy.call
+    context.client.cma.entries.get = spy.call
 
-    const task = initializeTask(applyAddEntitiesTask(), context)
+    context.changeset.items.push(updatedChangesetItem)
+
+    const task = initializeTask(ApplyChangesetTasks.createChangeEntitiesTask(), context)
     let error = null
     try {
       await task.run()
@@ -64,7 +63,7 @@ describe('applyAddEntitiesTask', () => {
     }
 
     expect(error).to.be.null
-    expect(task.tasks[0].output).to.be.equal('✨ successfully created added-entry')
-    expect(spy.called).to.be.equal(1)
+    expect(task.tasks[0].output).to.be.equal('✨ successfully updated update-entry')
+    expect(spy.called).to.be.equal(2)
   })
 })
