+++ test/unit/engine/apply-changeset/tasks/create-validate-changeset-task.test.ts
@@ -1,12 +1,12 @@
 import { initializeTask } from '../../../test-utils'
 import { expect } from 'chai'
-import { applyValidateChangesetTask } from '../../../../../src/engine/apply-changeset/tasks/create-validate-changeset-taskk'
 import { beforeEach } from 'mocha'
 import { AddedChangesetItem, UpdatedChangesetItem } from '../../../../../src/engine/types'
 import { createLinkObject } from '../../../../../src/engine/utils/create-link-object'
 import { createApplyChangesetContext } from '../../../fixtures/apply-changeset-context-fixture'
 import { ApplyChangesetContext } from '../../../../../src/engine/apply-changeset/types'
 import { ContainsMetadataError } from '../../../../../src/engine/errors'
+import { ApplyChangesetTasks } from '../../../../../src/engine/apply-changeset/tasks'
 
 describe('applyValidateChangesetTask', () => {
   let context: ApplyChangesetContext
@@ -35,7 +35,7 @@ describe('applyValidateChangesetTask', () => {
 
       context.changeset.items.push(updatedChangesetItem)
 
-      const task = initializeTask(applyValidateChangesetTask(), context)
+      const task = initializeTask(ApplyChangesetTasks.createValidateChangesetTask(), context)
       let error: ContainsMetadataError | null = null
       try {
         await task.run()
@@ -77,7 +77,7 @@ describe('applyValidateChangesetTask', () => {
 
       context.changeset.items.push(addedChangesetItem)
 
-      const task = initializeTask(applyValidateChangesetTask(), context)
+      const task = initializeTask(ApplyChangesetTasks.createValidateChangesetTask(), context)
       let error: ContainsMetadataError | null = null
       try {
         await task.run()
@@ -115,7 +115,7 @@ describe('applyValidateChangesetTask', () => {
 
       context.changeset.items.push(addedChangesetItem)
 
-      const task = initializeTask(applyValidateChangesetTask(), context)
+      const task = initializeTask(ApplyChangesetTasks.createValidateChangesetTask(), context)
       let error: ContainsMetadataError | null = null
       try {
         await task.run()
